+++ axum-extra/src/extract/mod.rs
@@ -1,8 +1,9 @@
 //! Additional extractors.
 
 mod cached;
+mod or_default;
 
-pub use self::cached::Cached;
+pub use self::{cached::Cached, or_default::OrDefault};
 
 pub mod rejection {
     //! Rejection response types.