+++ axum-extra/CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
-- None.
+- Add `OrDefault` extractor ([#589])
+
+[#589]: https://github.com/tokio-rs/axum/pull/589
 
 # 0.1.0 (02. December, 2021)